refactor(frontend): extract permission editing into separate component
