+++ src/i18n/locale/en.json
@@ -78,6 +78,32 @@
   "components.NotificationTypeSelector.mediafailedDescription": "Sends a notification when media fails to be added to services (Radarr/Sonarr). For certain agents, this will only send the notification to admins or users with the \"Manage Requests\" permission.",
   "components.NotificationTypeSelector.mediarequested": "Media Requested",
   "components.NotificationTypeSelector.mediarequestedDescription": "Sends a notification when new media is requested. For certain agents, this will only send the notification to admins or users with the \"Manage Requests\" permission.",
+  "components.PermissionEdit.admin": "Admin",
+  "components.PermissionEdit.adminDescription": "Full administrator access. Bypasses all permission checks.",
+  "components.PermissionEdit.advancedrequest": "Advanced Requests",
+  "components.PermissionEdit.advancedrequestDescription": "Grants permission to use advanced request options. (Ex. Changing servers/profiles/paths)",
+  "components.PermissionEdit.autoapprove": "Auto Approve",
+  "components.PermissionEdit.autoapproveDescription": "Grants auto approval for any requests made by this user.",
+  "components.PermissionEdit.autoapproveMovies": "Auto Approve Movies",
+  "components.PermissionEdit.autoapproveMoviesDescription": "Grants auto approve for movie requests made by this user.",
+  "components.PermissionEdit.autoapproveSeries": "Auto Approve Series",
+  "components.PermissionEdit.autoapproveSeriesDescription": "Grants auto approve for series requests made by this user.",
+  "components.PermissionEdit.managerequests": "Manage Requests",
+  "components.PermissionEdit.managerequestsDescription": "Grants permission to manage Overseerr requests. This includes approving and denying requests.",
+  "components.PermissionEdit.request": "Request",
+  "components.PermissionEdit.request4k": "Request 4K",
+  "components.PermissionEdit.request4kDescription": "Grants permission to request 4K movies and series.",
+  "components.PermissionEdit.request4kMovies": "Request 4K Movies",
+  "components.PermissionEdit.request4kMoviesDescription": "Grants permission to request 4K movies.",
+  "components.PermissionEdit.request4kTv": "Request 4K Series",
+  "components.PermissionEdit.request4kTvDescription": "Grants permission to request 4K Series.",
+  "components.PermissionEdit.requestDescription": "Grants permission to request movies and series.",
+  "components.PermissionEdit.settings": "Manage Settings",
+  "components.PermissionEdit.settingsDescription": "Grants permission to modify all Overseerr settings. A user must have this permission to grant it to others.",
+  "components.PermissionEdit.users": "Manage Users",
+  "components.PermissionEdit.usersDescription": "Grants permission to manage Overseerr users. Users with this permission cannot modify users with Administrator privilege, or grant it.",
+  "components.PermissionEdit.vote": "Vote",
+  "components.PermissionEdit.voteDescription": "Grants permission to vote on requests (voting not yet implemented)",
   "components.PersonDetails.appearsin": "Appears in",
   "components.PersonDetails.ascharacter": "as {character}",
   "components.PersonDetails.crewmember": "Crew Member",
@@ -455,41 +481,15 @@
   "components.TvDetails.userrating": "User Rating",
   "components.TvDetails.viewfullcrew": "View Full Crew",
   "components.TvDetails.watchtrailer": "Watch Trailer",
-  "components.UserEdit.admin": "Admin",
-  "components.UserEdit.adminDescription": "Full administrator access. Bypasses all permission checks.",
-  "components.UserEdit.advancedrequest": "Advanced Requests",
-  "components.UserEdit.advancedrequestDescription": "Grants permission to use advanced request options. (Ex. Changing servers/profiles/paths)",
-  "components.UserEdit.autoapprove": "Auto Approve",
-  "components.UserEdit.autoapproveDescription": "Grants auto approval for any requests made by this user.",
-  "components.UserEdit.autoapproveMovies": "Auto Approve Movies",
-  "components.UserEdit.autoapproveMoviesDescription": "Grants auto approve for movie requests made by this user.",
-  "components.UserEdit.autoapproveSeries": "Auto Approve Series",
-  "components.UserEdit.autoapproveSeriesDescription": "Grants auto approve for series requests made by this user.",
   "components.UserEdit.avatar": "Avatar",
   "components.UserEdit.edituser": "Edit User",
   "components.UserEdit.email": "Email",
-  "components.UserEdit.managerequests": "Manage Requests",
-  "components.UserEdit.managerequestsDescription": "Grants permission to manage Overseerr requests. This includes approving and denying requests.",
   "components.UserEdit.permissions": "Permissions",
-  "components.UserEdit.request": "Request",
-  "components.UserEdit.request4k": "Request 4K",
-  "components.UserEdit.request4kDescription": "Grants permission to request 4K movies and series.",
-  "components.UserEdit.request4kMovies": "Request 4K Movies",
-  "components.UserEdit.request4kMoviesDescription": "Grants permission to request 4K movies.",
-  "components.UserEdit.request4kTv": "Request 4K Series",
-  "components.UserEdit.request4kTvDescription": "Grants permission to request 4K Series.",
-  "components.UserEdit.requestDescription": "Grants permission to request movies and series.",
   "components.UserEdit.save": "Save",
   "components.UserEdit.saving": "Saving…",
-  "components.UserEdit.settings": "Manage Settings",
-  "components.UserEdit.settingsDescription": "Grants permission to modify all Overseerr settings. A user must have this permission to grant it to others.",
   "components.UserEdit.userfail": "Something went wrong saving the user.",
   "components.UserEdit.username": "Username",
-  "components.UserEdit.users": "Manage Users",
-  "components.UserEdit.usersDescription": "Grants permission to manage Overseerr users. Users with this permission cannot modify users with Administrator privilege, or grant it.",
   "components.UserEdit.usersaved": "User saved",
-  "components.UserEdit.vote": "Vote",
-  "components.UserEdit.voteDescription": "Grants permission to vote on requests (voting not yet implemented)",
   "components.UserList.admin": "Admin",
   "components.UserList.autogeneratepassword": "Automatically generate password",
   "components.UserList.create": "Create",
@@ -520,6 +520,7 @@
   "components.UserList.userdeleteerror": "Something went wrong deleting the user",
   "components.UserList.userlist": "User List",
   "components.UserList.username": "Username",
+  "components.UserList.userssaved": "Users saved",
   "components.UserList.usertype": "User Type",
   "components.UserList.validationemailrequired": "Must enter a valid email address.",
   "components.UserList.validationpasswordminchars": "Password is too short - should be 8 chars minimum.",
